survey-laurayinglu/requirements.txt
-i https://pypi.org/simple
authlib==1.2.0
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3 ; python_version >= '3.7'
cryptography==39.0.1 ; python_version >= '3.6'
flask==2.2.2
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
psycopg2-binary==2.9.5
pycparser==2.21
requests==2.28.2
setuptools==67.2.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2 ; python_version >= '3.7'
zipp==3.13.0 ; python_version >= '3.7'
